consolidate breakdown config types and impl with breakdown normalization
