+++ relay-general/src/protocol/mod.rs
@@ -31,9 +31,7 @@ mod user;
 mod user_report;
 
 pub use self::breadcrumb::Breadcrumb;
-pub use self::breakdowns::{
-    BreakdownConfig, Breakdowns, BreakdownsConfig, EmitBreakdowns, SpanOperationsConfig,
-};
+pub use self::breakdowns::Breakdowns;
 pub use self::clientsdk::{ClientSdkInfo, ClientSdkPackage};
 pub use self::constants::{INVALID_ENVIRONMENTS, INVALID_RELEASES, VALID_PLATFORMS};
 pub use self::contexts::{
